[MLIR] Shifted normalize-memrefs pass to lib/Transforms
-- The interprocedural pass to normalize memref layout maps is now shifted to lib/Transforms
-- Minor code fixes
-- A test case file for the pass has been added. More test cases to be added later.
